Refactor file-chooser to be a component.  Changed so when opening a file it opens into the monaco editor instead of expansion panel
