+++ src/app/app.module.ts
@@ -11,14 +11,13 @@ import { CovalentHighlightModule } from '@covalent/highlight';
 import { CovalentMarkdownModule } from '@covalent/markdown';
 import { CovalentChartsModule } from '@covalent/charts';
 import { CovalentMonacoEditorModule } from '../platform/monaco-editor';
-import { FileExpansionPanelComponent, TdContainerDirective } from '../components/';
+import { CovalentFileChooserModule } from '../platform/file-chooser';
+
 
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
-    FileExpansionPanelComponent,
-    TdContainerDirective,
   ], // directives, components, and pipes owned by this NgModule
   imports: [
     BrowserModule,
@@ -29,13 +28,11 @@ import { FileExpansionPanelComponent, TdContainerDirective } from '../components
     CovalentChartsModule.forRoot(),
     CovalentMarkdownModule.forRoot(),
     CovalentMonacoEditorModule.forRoot(),
+    CovalentFileChooserModule.forRoot(),
     appRoutes,
   ], // modules needed to run this module
   providers: [
     appRoutingProviders,
-  ], // additional providers needed for this module
-  entryComponents: [
-    FileExpansionPanelComponent,
   ],
   bootstrap: [ AppComponent ],
 })
