+++ src/app/components/home/home.component.html
@@ -11,16 +11,7 @@
                             <button md-icon-button><md-icon>more_vert</md-icon></button>
                         </div>
                         <div flex layout="column" class="md-content">
-                            <md-nav-list dense>
-                                <h3 md-subheader>Browse Files</h3>
-                                <template let-file let-last="last" ngFor [ngForOf]="files">
-                                    <a md-list-item class="directory" (click)="loadFiles(file.name, expansionContainer.viewContainerRef)">
-                                      <md-icon md-list-icon>{{file.icon}}</md-icon> {{file.name}}
-                                    </a>
-                                    <div #expansionContainer="tdContainer" tdContainer></div>
-                                    <md-divider *ngIf="!last"></md-divider>
-                                </template>
-                            </md-nav-list>
+                            <td-file-chooser (openFile)="monaco.value = $event"></td-file-chooser>
                         </div>
                     </div>
                     <div class="td-layout-footer bgc-black pad-xs" layout="row">
