+++ src/platform/file-chooser/file-chooser.component.html
@@ -1,0 +1,9 @@
+<md-nav-list dense>
+    <h3 md-subheader>Browse Files</h3>
+    <template let-file let-last="last" ngFor [ngForOf]="files">
+        <a md-list-item class="directory" (click)="loadFiles(file.name)">
+            <md-icon md-list-icon>{{file.icon}}</md-icon> {{file.name}}
+        </a>
+        <md-divider *ngIf="!last"></md-divider>
+    </template>
+</md-nav-list>
