+++ src/platform/file-chooser/file-chooser.component.scss
@@ -1,0 +1,3 @@
+:host {
+    display:block;
+}
