+++ src/platform/file-chooser/index.ts
@@ -1,0 +1,2 @@
+export { TdFileChooserComponent } from './file-chooser.component';
+export { CovalentFileChooserModule } from './file-chooser.module';
