+++ src/platform/file-chooser/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@covalent/monaco-editor",
+  "version": "0.9.1",
+  "description": "Teradata UI Platform Monaco Editor Module",
+  "keywords": [
+    "angular",
+    "components",
+    "reusable",
+    "monaco",
+    "editor",
+    "browse"
+  ],
+  "scripts": {
+  },
+  "engines": {
+    "node": ">4.4 < 7",
+    "npm": ">= 3"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/teradata/covalent.git"
+  },
+  "bugs": {
+    "url": "https://github.com/teradata/covalent/issues"
+  },
+  "license": "MIT",
+  "author": "Teradata UX",
+  "contributors": [
+    "Kyle Ledbetter <kyle.ledbetter@teradata.com>",
+    "Richa Vyas <richa.vyas@teradata.com>",
+    "Ed Morales <eduardo.morales@teradata.com>",
+    "Jason Weaver <jason.weaver@teradata.com>",
+    "Jeremy Wilken <jeremy.wilken@teradata.com>",
+    "Jeremy Smartt <jeremy.smartt@teradata.com>"
+  ],
+  "dependencies": {
+    "@covalent/core": "0.10.2"
+  }
+}
